+++ lib/map/tiled/world_map_by_tiled.dart
@@ -1,11 +1,12 @@
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/tiled/builder/tiled_world_builder.dart';
+import 'package:bonfire/map/tiled/builder/tiled_world_builder.dart';
 import 'package:flutter/widgets.dart';
+import 'package:tiledjsonreader/map/tiled_map.dart';
 
 class WorldMapByTiled extends WorldMap {
   late TiledWorldBuilder _builder;
   WorldMapByTiled(
-    TiledReader reader, {
+    WorldMapReader<TiledMap> reader, {
     Vector2? forceTileSize,
     ValueChanged<Object>? onError,
     double sizeToUpdate = 0,